change Calculator.java add double.; change CalculatorTest.java changed checkout procedure

+++ Lesson_2/Calculator.java
@@ -54,7 +54,7 @@ public class Calculator {
                 System.out.println(x * y);
                 break;
             case "/" : 
-                System.out.println(x / y);
+                System.out.println((double)x / y);
                 break; 
             case "^" : 
                 result = 1;
